requirements.txt
python-telegram-bot==20.7
transformers==4.36.2
soundfile==0.12.1
flask==3.0.0
flask-cors==4.0.0 